requirements.txt
Flask
Flask-Login
Flask-WTF
pyModbusTCP
Flask-SQLAlchemy
flask-caching
pymysql
cryptography
python-dotenv
apscheduler
requests
pytz